+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Adds
 
-- Adds the `FilterBarPopup` compoonent.
+- Adds the `FilterBarPopup` compoonent and `useFilterBarPopup` hook.
 
 ## 3.1.7 (July 3, 2024)